+++ ElvUI_MerathilisUI/Core/Changelog/6.61.lua
@@ -1,8 +1,15 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[661] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "23.04.2025",
+	FIXES = {
+		"[Core]: Changes due to 11.1.5",
+		"[Combat Alert]: Fix backdrop",
+	},
+	NEW = {
+		"[Notification]: Add an alert for full bags",
+	},
+	IMPROVEMENTS = {
+		"[GameMenu]: Some more info's on the GameMenu",
+	},
 }