+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ### Changes ###
 
-  * [Install]: Fix an issue with the last version where i made a whoopsi
+  * [TradeFrame]: Blizzard made the InputMoneyFrame secure, so no AddOn can change it anymore. Looks ugly, but hey
+  * [Core]: Use xpcall to grab some more errors for the init of the Modules
+  * [TradeTabs]: Display by default Herbalism and Tailoring tab
+  * [CollectionJournal]: Fixed on of the Tabs at the bottom