+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changes ###
 
-  * [Armory]: More settings and visual improvements
-  * [GameMenu]: Some more infos on the GameMenu
-  * [GameMenu]: Fix Option -fix1
+  * [Core]: Changes due to 11.1.5
+  * [Combat Alert]: Fix backdrop
+  * [Notification]: Add an alert for full bags
+  * [GameMenu]: Some more info's on the GameMenu